Register navigator to entities instead of views
