+++ airgun/views/architecture.py
@@ -1,7 +1,5 @@
-from navmazing import NavigateToSibling
 from widgetastic.widget import ParametrizedView, View, Text, TextInput
 
-from airgun.navigation import BaseNavigator, menu_click, navigator
 from airgun.widgets import ResourceList, Search
 
 
@@ -37,26 +35,3 @@ class ArchitectureDetailsView(ParametrizedView):
 
     def submit_data(self):
         self.browser.click(self.submit)
-
-
-@navigator.register(ArchitectureView, 'All')
-class ShowAllArchitectures(BaseNavigator):
-    VIEW = ArchitectureView
-
-    def step(self, *args, **kwargs):
-        menu_click(
-            ["//a[@id='hosts_menu']", self.view.navigate_locator],
-            self.view.browser
-        )
-
-
-@navigator.register(ArchitectureView, 'New')
-class AddNewArchitecture(BaseNavigator):
-    VIEW = ArchitectureDetailsView
-
-    prerequisite = NavigateToSibling('All')
-
-    def step(self, *args, **kwargs):
-        self.view.browser.wait_for_element(
-            self.parent.new, ensure_page_safe=True)
-        self.parent.browser.click(self.parent.new)
